Correct username on Minecraft flags
When it's a Minecraft player, it's a webhook user. This webhook user changes username, and thus the mention changes, every time a new messages is sent in the MC chat.
That means that this flag will show the wrong user if it's a minecraft player that flagged, and another player and talked since.

This PR should fix that.